libemu: implement push/pop dword ptr fs:[0x0]
The emulator already set the segment to FS, which is correct, but it
also pushes the values to the FS segment which is causes emulator
errors due to invalid reads.
So explicitly set back the memory segment to SS during the push.
This issue doesn't show up for other segments, since FS is the only one which is
a different segment (from the emulator's p.o.v.).

Signed-off-by: Török Edwin <edwintorok@gmail.com>
Signed-off-by: common <common@carnivore.it>