Remove harmful rpath in LDFLAGS and fix logic for cargo. This patch is from Debian libemu package - libemu-03_remove_rpath_and_fix_ldflags.patch https://packages.debian.org/search?searchon=sourcenames&keywords=libemu
Description: Remove harmful rpath in LDFLAGS and fix logic for cargo.
Before this patch, libemu was searching for includes and also had an rpath
for libraries in /usr/local/lib. Also the logic for libcargo is wrong, as
makes the content of LDFLAGS empty if cargo is enabled but there are no
headers in the system. This was preventing proper hardening to be activated
in Debian.
Author: David Martínez Moreno ender@debian.org
Forwarded: no
Reviewed-By: David Martínez Moreno ender@debian.org
Last-Update: 2012-10-14